# Try to call Less() fewer times when iterating