--- a/btree.go
+++ b/btree.go
@@ -504,13 +504,18 @@ const (
 // "greaterThan" or "lessThan" queries.
 func (n *node) iterate(dir direction, start, stop Item, includeStart bool, hit bool, iter ItemIterator) (bool, bool) {
 	var ok, found bool
-	var index int
+	var index, stopIndex int
 	switch dir {
 	case ascend:
 		if start != nil {
 			index, _ = n.items.find(start)
 		}
-		for i := index; i < len(n.items); i++ {
+		if stop != nil {
+			stopIndex, _ = n.items.find(stop)
+		} else {
+			stopIndex = len(n.items)
+		}
+		for i := index; i < stopIndex; i++ {
 			if len(n.children) > 0 {
 				if hit, ok = n.children[i].iterate(dir, start, stop, includeStart, hit, iter); !ok {
 					return hit, false
@@ -521,15 +526,12 @@ func (n *node) iterate(dir direction, start, stop Item, includeStart bool, hit b
 				continue
 			}
 			hit = true
-			if stop != nil && !n.items[i].Less(stop) {
-				return hit, false
-			}
 			if !iter(n.items[i]) {
 				return hit, false
 			}
 		}
 		if len(n.children) > 0 {
-			if hit, ok = n.children[len(n.children)-1].iterate(dir, start, stop, includeStart, hit, iter); !ok {
+			if hit, ok = n.children[stopIndex].iterate(dir, start, stop, includeStart, hit, iter); !ok {
 				return hit, false
 			}
 		}
@@ -542,6 +544,16 @@ func (n *node) iterate(dir direction, start, stop Item, includeStart bool, hit b
 		} else {
 			index = len(n.items) - 1
 		}
+
+		if stop != nil {
+			stopIndex, found = n.items.find(stop)
+			if found {
+				stopIndex += 1
+			}
+		} else {
+			stopIndex = 0
+		}
+
 		for i := index; i >= 0; i-- {
 			if start != nil && !n.items[i].Less(start) {
 				if !includeStart || hit || start.Less(n.items[i]) {
@@ -562,7 +574,7 @@ func (n *node) iterate(dir direction, start, stop Item, includeStart bool, hit b
 			}
 		}
 		if len(n.children) > 0 {
-			if hit, ok = n.children[0].iterate(dir, start, stop, includeStart, hit, iter); !ok {
+			if hit, ok = n.children[stopIndex].iterate(dir, start, stop, includeStart, hit, iter); !ok {
 				return hit, false
 			}
 		}