--- a/btree_generic.go
+++ b/btree_generic.go
@@ -499,13 +499,18 @@ func empty[T any]() optionalItem[T] {
 // "greaterThan" or "lessThan" queries.
 func (n *node[T]) iterate(dir direction, start, stop optionalItem[T], includeStart bool, hit bool, iter ItemIteratorG[T]) (bool, bool) {
 	var ok, found bool
-	var index int
+	var index, stopIndex int
 	switch dir {
 	case ascend:
 		if start.valid {
 			index, _ = n.items.find(start.item, n.cow.less)
 		}
-		for i := index; i < len(n.items); i++ {
+		if stop.valid {
+			stopIndex, _ = n.items.find(stop.item, n.cow.less)
+		} else {
+			stopIndex = len(n.items)
+		}
+		for i := index; i < stopIndex; i++ {
 			if len(n.children) > 0 {
 				if hit, ok = n.children[i].iterate(dir, start, stop, includeStart, hit, iter); !ok {
 					return hit, false
@@ -516,15 +521,12 @@ func (n *node[T]) iterate(dir direction, start, stop optionalItem[T], includeSta
 				continue
 			}
 			hit = true
-			if stop.valid && !n.cow.less(n.items[i], stop.item) {
-				return hit, false
-			}
 			if !iter(n.items[i]) {
 				return hit, false
 			}
 		}
 		if len(n.children) > 0 {
-			if hit, ok = n.children[len(n.children)-1].iterate(dir, start, stop, includeStart, hit, iter); !ok {
+			if hit, ok = n.children[stopIndex].iterate(dir, start, stop, includeStart, hit, iter); !ok {
 				return hit, false
 			}
 		}
@@ -537,7 +539,15 @@ func (n *node[T]) iterate(dir direction, start, stop optionalItem[T], includeSta
 		} else {
 			index = len(n.items) - 1
 		}
-		for i := index; i >= 0; i-- {
+		if stop.valid {
+			stopIndex, found = n.items.find(stop.item, n.cow.less)
+			if found {
+				stopIndex += 1
+			}
+		} else {
+			stopIndex = 0
+		}
+		for i := index; i >= stopIndex; i-- {
 			if start.valid && !n.cow.less(n.items[i], start.item) {
 				if !includeStart || hit || n.cow.less(start.item, n.items[i]) {
 					continue
@@ -548,16 +558,13 @@ func (n *node[T]) iterate(dir direction, start, stop optionalItem[T], includeSta
 					return hit, false
 				}
 			}
-			if stop.valid && !n.cow.less(stop.item, n.items[i]) {
-				return hit, false //	continue
-			}
 			hit = true
 			if !iter(n.items[i]) {
 				return hit, false
 			}
 		}
 		if len(n.children) > 0 {
-			if hit, ok = n.children[0].iterate(dir, start, stop, includeStart, hit, iter); !ok {
+			if hit, ok = n.children[stopIndex].iterate(dir, start, stop, includeStart, hit, iter); !ok {
 				return hit, false
 			}
 		}
@@ -901,7 +908,10 @@ func (a Int) Less(b Item) bool {
 // goroutines, but Read operations are.
 type BTree BTreeG[Item]
 
+var LessCount = 0
+
 var itemLess LessFunc[Item] = func(a, b Item) bool {
+	LessCount++
 	return a.Less(b)
 }